zsh/_meson: update the completion for `meson subprojects`
2 new subcommands:
* purge
* packagefiles

Also, there seems to be no `status` subcommand.

Turns out the entire "${${commands[(r)name:*]}%%:*}" gadget is not
needed yet. It can return in the future if the subcommand chain would
not be uniquely determined by the function that produces its completion;
then we would have to fetch the subcommand names from the array.