dependencies/detect: correctly get the `log_tried()`
We're currently trying to call the method `.log_tried()` on a function,
which is invalid. Mypy even helpfully points this out, except that we
don't yet run mypy on this file :/

We can fix this, but we may not be able to figure out the method anyway,
since we could end up with a situation where there is no way to get the
method tried, since the dependency errored on creation and the callable
isn't an `ExternalDependency` helper.

Fixes: #11145