Fix find_program pkg-config when defined in machine file
When pkg-config is defined in the [binaries] section of a machine
file, the entry is called 'pkgconfig', so we should match that when
the user does find_program('pkg-config').

This is usually not needed because the dependency handler takes care
of it, but sometimes the user wants to access pkg-config directly;
for example, to pass to external build systems like cargo.