fix build_target(objects: ...) documentation
The documentation for build_target(...) does not list file or str as
the possible types for the "objects" keyword argument, even though in
theory the argument is meant for prebuild object files that are part
of the sources.

Of course that is only the theory, because an ExtractedObjects object
is probably used a lot more than a file in the source tree.  But
at least make the reference manual's typing information accurate.