Revert "mtest: handle TAP tests with unknown version."
This reverts commit b7a5c384a1f1ba80c09904e7ef4f5160bdae3345.

The added warning triggers widely, is not actually required by the TAP spec,
even version 14. Even if it were required as of TAP 14, it'd amount to
suddenly requiring TAP 14, given that the version number previously was not
required.

The warning causes trouble for e.g. PostgreSQL due to its verbosity.