Allow `true` and `false` as values to feature options
Currently, if an option is a boolean, and you want to upgrade it to a
feature in a backwards compatible way, you need to first convert it to a
combo option with `true` and `false` deprecated, set on that for a
while, then change it to a feature. There really isn't a reason for that
though, we can just accept `true` and `false` as values to features, and
then convert them to `enabled` and `disabled`, respectively.