type_checking: add feature_validator checks to replace since_values
and deprecated_values.

Both of these are oddly specific to be in `typed_kwargs` directly (my
fault), since they have very limited use. They've also been extended
past what I'd originally envisioned, which was a simple "check for
elements in arrays and dictionaries" to include checking for types. This
bends them and makes them complicated and fragile.

Instead, we can leverage the feature_validator to handle these, this
reduces the complexity of `typed_kwargs`, by essentially splitting the
`emit_feature_change` function out of that, and into it's own separate
function. Though, on that topic, it's actually now two separate, but
related functions. One that checks for types, and one that checks for
values.

This matters a lot when we're checking inside of containers,
particularly dictionaries. Because the original code meant to look at
keys, which makes sense when checking for values `x in container`, but
for types that doesn't make sense, we need to look at the values, to see
if there are any values that are of the wrong time (Meson dictionary
keys are always strings).