gnome: support using generate_gir from gobject-introspection
From gobject-introspection itself the use of gnome.generate_gir is
not possible leading to several hacks that are hard to maintain in
order to support building gobject-introspection with submodules.

This change in combination with https://gitlab.gnome.org/GNOME/gobject-introspection/-/merge_requests/380
uses a new `gobject-introspection-no-typelibs` dependency to allow
defining dependency to allow using the GI dependency prior to having
the typelibs generated.