compilers/detect: use detect_compiler_for instead of detect_X_compiler
The former looks in the cached compiler list, while the latter doesn't.
This means that if we have already detected the compiler, it will be
re-detected, going through expensive lookup paths. Doing this also adds
the newly detected compiler to the cache, avoiding the need to look it up
again