envconfig: Update MachineInfo inplace when redetecting cpu info
Currently we replace these with new instances. This is nice from a
simplicity point of view. The problem is that we store references to
the MachineInfo instance, instead of looking it up fresh in the
environment. Because of that we can end up with stale references. By
mutating in place we avoid that.