Avoid replacing symlinks when installing with --only-changed
We are using install --only-changed while other code is concurrently
traversing an unchanged part of the tree (a python import), but even
though this should be fine it sometimes fails as an unchanged symlink
can disappear. To avoid, check if the symlink actually needs to be
changed, otherwise leave it as is.