CI: migrate pypy3 tests from bionic to Arch
The bionic image is really old and mainly exists to test that Meson
itself still works on really old distros (and really old python).
Ideally we'd avoid depending too much on it.

We can get a very modern pypy3 automatically this way, and potentially
use it for more stuff too.