mdevenv: Add --relocate option
It replaces any path that starts with the root source directory with a
new location.

This is useful when running the project on a different machine, for
example when the tree is on an embedded device but mounted with NFS on a
build machine. In that case, the source tree could be on a device at
`/home/user/myproject` and it gets mounted on a build machine to
`/nfs/myproject`. Environment would contain variables like
`PATH=/nfs/myproject/builddir`, running on the device
`meson devenv --relocate /home/user/myproject` would replace that with
`PATH=/home/user/myproject/builddir`.