work around circular imports in python probe script
It seems this happens because at some point setuptools imports gettext,
and we have a script by the same name.

In general, this path injection by default is bad news for our use case.
Python 3.11 introduced -P for this reason, but we cannot depend on that.
Instead, check for it first, and delete it, before doing more imports.