interpreter: deprecate lang_args an and lang_link_args in project()
It's a trap, because it doesn't do what anyone wants, to put language
args in the project() default_options. The expectation is that these
will always be applied, what actually happens is that the environment
variable is not read, machine files are not read, and the options are
replaced if -D<lang>_args is passed on the command line. No one wants
this, they really want to use `add_project_aruments` or
`add_global_arguments`.