Revert "Exit meson with an error if an invalid escape sequence is found in a"
This reverts commit 348248f0a19bdc80e8a184befb2faaa1d5e66f40.

The rules were relaxed in commit ccc4ce28cc9077d77a0bc9e72b1177eba1be7186
to permit this, so it's never possible to raise this exception anymore.
But that commit was incomplete, and didn't remove the now-useless
infrastructure for exception handling.

The test needed to test this was always broken, and then removed in
commit 465ef856ac9b978f13414db4aff649c66f2e6be5, and still this useless
try/except persisted.