build|interpreter: allow build_target.extra_files to be CustomTarge and indexes
We previously documented that `CustomTarget` and `CustomTargetIndex`
were allowed. But this wasn't true, as there was an assertion that we
got a `File`. To maintain this behavior we set the allowed types to
`File | str`. It seems useful to allow this, however. So this patch adds
that back, though as a new feature for 1.1.0. Since it never worked to
do so before, this seems like the correct approach

Fixes: #11512