interpreter: validate build_target(rust_crate_type)
I need to do some more review of this to make sure that the we haven't
stopped allowing something that worked, even if it was undesirable.

This allows us to clean up a ton of code that has to handle the fact
that rust_crate_type wasn't always defined before this. It now is and
thus all of the `hasattr()` checks can go away