backends: fix bug where meson_exe crashed if constructed with found programs
Because we base the pickled data name on the name property of the
command being run... and for built targets, `exe.name` is always just
the name. However, for an ExternalProgram this is just whatever string
we searched for, so, NOT just the basename.

This became a bigger issue once we started using generator() with the
actual program in commit 6aeec808367f05463394e30a8d40834e97c7afc0,
rather than first casting it to a string, because the VS backend
*always* uses the meson_exe approach for various reasons related to VS
being VS.

Outside of that, it's difficult to actually get an ExternalProgram
object passed to meson_exe -- CustomTarget lowers it to a string,
capture is handled via argparse instead of pickling, etc.

Fixes #11593