tests: fix regression that broke mtest_inprocess
In commit faf79f4539841cbf89fe8d53cf35aa91fd8273c9 we broke this utility
function by referencing a non-existent variable. Fortunately, the only
time we ever used said function was once, and that in a test case where
we tested that it raised a MesonException before the undefined variable
error could occur.